+++ data/configs/Bolshoi.config
@@ -4,6 +4,6 @@ L = 250
 Epsilon = 1
 Mp = 1.36297e+08
 MinParticles = 50
-NameIndex = /home/imendoza/nbody-relaxed/params/name_index.txt
-TypeIndex = /home/imendoza/nbody-relaxed/params/full_type_index.txt
+NameIndex = /home/imendoza/nbody-relaxed/data/configs/name_index.txt
+TypeIndex = /home/imendoza/nbody-relaxed/data/configs/full_type_index.txt
