+++ data/configs/BolshoiP.config
@@ -4,6 +4,6 @@ L = 250
 Epsilon = 1
 Mp = 1.54975e+08
 MinParticles = 50
-NameIndex = name_index.txt
-TypeIndex = full_type_index.txt
+NameIndex = /home/imendoza/nbody-relaxed/data/configs/name_index.txt
+TypeIndex = /home/imendoza/nbody-relaxed/data/configs/full_type_index.txt
