--- data/configs/all_vars.txt
@@ -1,1 +1,0 @@
-x y z vx vy vz jx jy jz ax ay az ax(500c) ay(500c) az(500c) gamma_log_vmax_tdyn gamma_log_vmax_inst log_(vmax/vmax_max(tdyn;tmpeak)) gamma_tdyn gamma_100myr gamma_2tdyn gamma_3tdyn gamma_4tdyn gamma_inst gamma_mpeak snap_num snap_idx scale halfmass_scale acc_scale m4%_scale mpeak_scale first_acc_scale scale_of_last_mm b_to_a b_to_a(500c) c_to_a c_to_a(500c) breadth_first_id depth_first_id desc_id desc_pid desc_scale descid last_mainleaf_depthfirst_id last_progenitor_depthfirst_id next_coprogenitor_depthfirst_id orig_halo_id tree_root_id future_merger_mmp_id time_to_future_merger first_acc_mvir first_acc_vmax id pid upid mmp? np num_prog phantom mvir  mvir_all sam_mvir m200b m200c m500c m2500c m_pe_behroozi m_pe_diemer macc mpeak rvir rs rs_klypin rvmax halfmass_radius spin spin_bullock t/|u| voff xoff tidal_force tidal_force_tdyn tidal_id vmax vpeak vacc vmax@mpeak vrms
